refactor(log): move Log into its own DDD package
Completes the project-wide move to package-by-feature/DDD structure: Log
and LogRepository move from domain/ into log/, mirroring the other feature
packages. No behavior change; domain/, controller/, and service/ (the old
layer-based packages) are now empty.
